build(pom): add errorprone analyzer
errorprone is used to analyze code without remote analysis.

FieldCanBeLocal is disabled right now because most of the code is
unused at the moment.

ImmutableMemberCollection is disabled right now because guava isn't
strictly needed at this moment.

Java7ApiChecker is disabled for obvious reasons - we can use Java8 apis.

IdentifierName is disabled because good heavens there are a lot of
acronyms.

Varifier is disabled because Java 8 does not support "var".